Retrieve the gene sets data frame

Retrieve a data frame of gene sets and their member genes. The available species and collections can be checked with msigdbr_species() and msigdbr_collections().

msigdbr(species = "Homo sapiens", category = NULL, subcategory = NULL)

Arguments

species
Species name, such as Homo sapiens or Mus musculus.

category
MSigDB collection abbreviation, such as H or C1.

subcategory
MSigDB sub-collection abbreviation, such as CGP or BP.

Value

A data frame of gene sets with one gene per row.

References

https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp

Examples

# get all human gene sets
# \donttest{
msigdbr(species = "Homo sapiens")

#> # A tibble: 3,862,192 x 18
#>    gs_cat gs_subcat      gs_name        entrez_gene ensembl_gene    gene_symbol
#>    <chr>  <chr>          <chr>                <int> <chr>           <chr>      
#>  1 C3     MIR:MIR_Legacy AAACCAC_MIR140       10257 ENSG00000125257 ABCC4      
#>  2 C3     MIR:MIR_Legacy AAACCAC_MIR140       23172 ENSG00000165660 ABRAXAS2   
#>  3 C3     MIR:MIR_Legacy AAACCAC_MIR140          81 ENSG00000130402 ACTN4      
#>  4 C3     MIR:MIR_Legacy AAACCAC_MIR140          90 ENSG00000115170 ACVR1      
#>  5 C3     MIR:MIR_Legacy AAACCAC_MIR140        8754 ENSG00000168615 ADAM9      
#>  6 C3     MIR:MIR_Legacy AAACCAC_MIR140       11096 ENSG00000154736 ADAMTS5    
#>  7 C3     MIR:MIR_Legacy AAACCAC_MIR140      219287 ENSG00000165566 AMER2      
#>  8 C3     MIR:MIR_Legacy AAACCAC_MIR140         287 ENSG00000145362 ANK2       
#>  9 C3     MIR:MIR_Legacy AAACCAC_MIR140        8539 ENSG00000166181 API5       
#> 10 C3     MIR:MIR_Legacy AAACCAC_MIR140         571 ENSG00000156273 BACH1      
#> # … with 3,862,182 more rows, and 12 more variables: human_entrez_gene <int>,
#> #   human_gene_symbol <chr>, gs_id <chr>, gs_pmid <chr>, gs_geoid <chr>,
#> #   gs_exact_source <chr>, gs_url <chr>, gs_description <chr>,
#> #   species_name <chr>, species_common_name <chr>, ortholog_sources <chr>,
#> #   num_ortholog_sources <dbl>

# }

# get mouse C2 (curated) CGP (chemical and genetic perturbations) gene sets
# \donttest{
msigdbr(species = "Mus musculus", category = "C2", subcategory = "CGP")

#> # A tibble: 366,082 x 18
#>    gs_cat gs_subcat gs_name             entrez_gene ensembl_gene     gene_symbol
#>    <chr>  <chr>     <chr>                     <int> <chr>            <chr>      
#>  1 C2     CGP       ABBUD_LIF_SIGNALIN…       66395 ENSMUSG00000069… Ahnak      
#>  2 C2     CGP       ABBUD_LIF_SIGNALIN…       11658 ENSMUSG00000022… Alcam      
#>  3 C2     CGP       ABBUD_LIF_SIGNALIN…       71452 ENSMUSG00000020… Ankrd40    
#>  4 C2     CGP       ABBUD_LIF_SIGNALIN…       93760 ENSMUSG00000007… Arid1a     
#>  5 C2     CGP       ABBUD_LIF_SIGNALIN…       12040 ENSMUSG00000032… Bckdhb     
#>  6 C2     CGP       ABBUD_LIF_SIGNALIN…      239691 ENSMUSG00000051… AU021092   
#>  7 C2     CGP       ABBUD_LIF_SIGNALIN…       73647 ENSMUSG00000031… Capn9      
#>  8 C2     CGP       ABBUD_LIF_SIGNALIN…       12484 ENSMUSG00000047… Cd24a      
#>  9 C2     CGP       ABBUD_LIF_SIGNALIN…       20430 ENSMUSG00000030… Cyfip1     
#> 10 C2     CGP       ABBUD_LIF_SIGNALIN…       28199 ENSMUSG00000022… Dcaf11     
#> # … with 366,072 more rows, and 12 more variables: human_entrez_gene <int>,
#> #   human_gene_symbol <chr>, gs_id <chr>, gs_pmid <chr>, gs_geoid <chr>,
#> #   gs_exact_source <chr>, gs_url <chr>, gs_description <chr>,
#> #   species_name <chr>, species_common_name <chr>, ortholog_sources <chr>,
#> #   num_ortholog_sources <dbl>

# }