msigdbr: MSigDB Gene Sets for Multiple Organisms in a Tidy Data Format

Provides the 'Molecular Signatures Database' (MSigDB) gene sets typically used with the 'Gene Set Enrichment Analysis' (GSEA) software (Subramanian et al. 2005 doi:10.1073/pnas.0506580102 , Liberzon et al. 2015 doi:10.1016/j.cels.2015.12.004 , Castanza et al. 2023 doi:10.1038/s41592-023-02014-7 ) as an R data frame. The package includes the human genes as listed in MSigDB as well as the corresponding symbols and IDs for frequently studied model organisms such as mouse, rat, pig, fly, and yeast.

See also

Author

Maintainer: Igor Dolgalev igor.dolgalev@nyumc.org (ORCID)