Retrieve the gene sets data frame

Retrieve a data frame of gene sets and their member genes. The original human genes can be converted into their corresponding counterparts in various model organisms, including mouse, rat, pig, zebrafish, fly, and yeast. The output includes gene symbols along with NCBI and Ensembl IDs.

Usage

msigdbr(
  species = "Homo sapiens",
  db_species = "HS",
  collection = NULL,
  subcollection = NULL,
  category = deprecated(),
  subcategory = deprecated()
)

Arguments

species

Species name for output genes, such as "Homo sapiens" or "Mus musculus". Use msigdbr_species() for available options.

db_species

Species abbreviation for the human or mouse databases ("HS" or "MM").

collection

Collection abbreviation, such as "H" or "C1". Use msigdbr_collections() for the available options.

subcollection

Sub-collection abbreviation, such as "CGP" or "BP". Use msigdbr_collections() for the available options.

category

use the collection argument

subcategory

use the subcollection argument

Value

A tibble (a data frame with class tibble::tbl_df) of gene sets with one gene per row.

Details

Historically, the MSigDB resource has been tailored to the analysis of human-specific datasets, with gene sets exclusively aligned to the human genome. Starting with release 2022.1, MSigDB incorporated a database of mouse-native gene sets and was split into human and mouse divisions ("Hs" and "Mm"). Each one is provided in the approved gene symbols of its respective species. The versioning convention of MSigDB is in the format Year.Release.Species. The genes within each gene set may originate from a species different from the database target species as indicated by the gs_source_species and db_target_species fields.

Mouse MSigDB includes gene sets curated from mouse-centric datasets and specified in native mouse gene identifiers, eliminating the need for ortholog mapping.

References

https://www.gsea-msigdb.org/gsea/msigdb/index.jsp