Retrieve the gene sets data frame

Retrieve a data frame of gene sets and their member genes. The available species and collections can be checked with msigdbr_species() and msigdbr_collections().

msigdbr(species = "Homo sapiens", category = NULL, subcategory = NULL)

Arguments

species
Species name, such as Homo sapiens or Mus musculus.

category
MSigDB collection abbreviation, such as H or C1.

subcategory
MSigDB sub-collection abbreviation, such as CGP or BP.

Value

A data frame of gene sets with one gene per row.

References

https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp

Examples

# get all human gene sets
# \donttest{
msigdbr(species = "Homo sapiens")

#> # A tibble: 3,647,311 x 17
#>    gs_cat gs_subcat gs_name entrez_gene gene_symbol human_entrez_ge…
#>    <chr>  <chr>     <chr>         <int> <chr>                  <int>
#>  1 C3     MIR:MIR_… AAACCA…       10257 ABCC4                  10257
#>  2 C3     MIR:MIR_… AAACCA…       23172 ABRAXAS2               23172
#>  3 C3     MIR:MIR_… AAACCA…          81 ACTN4                     81
#>  4 C3     MIR:MIR_… AAACCA…          90 ACVR1                     90
#>  5 C3     MIR:MIR_… AAACCA…        8754 ADAM9                   8754
#>  6 C3     MIR:MIR_… AAACCA…       11096 ADAMTS5                11096
#>  7 C3     MIR:MIR_… AAACCA…      219287 AMER2                 219287
#>  8 C3     MIR:MIR_… AAACCA…         287 ANK2                     287
#>  9 C3     MIR:MIR_… AAACCA…        8539 API5                    8539
#> 10 C3     MIR:MIR_… AAACCA…         571 BACH1                    571
#> # … with 3,647,301 more rows, and 11 more variables: human_gene_symbol <chr>,
#> #   gs_id <chr>, gs_pmid <chr>, gs_geoid <chr>, gs_exact_source <chr>,
#> #   gs_url <chr>, gs_description <chr>, species_name <chr>,
#> #   species_common_name <chr>, ortholog_sources <chr>,
#> #   num_ortholog_sources <dbl>

# }

# get mouse C2 (curated) CGP (chemical and genetic perturbations) gene sets
# \donttest{
msigdbr(species = "Mus musculus", category = "C2", subcategory = "CGP")

#> # A tibble: 363,345 x 17
#>    gs_cat gs_subcat gs_name entrez_gene gene_symbol human_entrez_ge…
#>    <chr>  <chr>     <chr>         <int> <chr>                  <int>
#>  1 C2     CGP       ABBUD_…       66395 Ahnak                  79026
#>  2 C2     CGP       ABBUD_…       11658 Alcam                    214
#>  3 C2     CGP       ABBUD_…       71452 Ankrd40                91369
#>  4 C2     CGP       ABBUD_…       93760 Arid1a                  8289
#>  5 C2     CGP       ABBUD_…       12040 Bckdhb                   594
#>  6 C2     CGP       ABBUD_…      239691 AU021092              146556
#>  7 C2     CGP       ABBUD_…       73647 Capn9                  10753
#>  8 C2     CGP       ABBUD_…       12484 Cd24a              100133941
#>  9 C2     CGP       ABBUD_…       77644 C330007P06…            63932
#> 10 C2     CGP       ABBUD_…       20430 Cyfip1                 23191
#> # … with 363,335 more rows, and 11 more variables: human_gene_symbol <chr>,
#> #   gs_id <chr>, gs_pmid <chr>, gs_geoid <chr>, gs_exact_source <chr>,
#> #   gs_url <chr>, gs_description <chr>, species_name <chr>,
#> #   species_common_name <chr>, ortholog_sources <chr>,
#> #   num_ortholog_sources <dbl>

# }